Add functionality to remove riders
This functionality does not delete users overall, just unlinks them
from a ride.

Functionality to add/remove riders from a ride is linked to the
view-ride page.

There are a few odd behaviors w/ the lynx browser that could be
hardened, but that can be addressed later. It's minimally functional
as is.